# Improve performance of yarn build:clean