… AGENTS.md
… packages/AGENTS.md