## Motivation

Improve performance of the `yarn build:clean` command to reduce build times and enhance developer productivity.

## Changes

- **Performance optimisation**: Enhanced the `yarn build:clean` command to execute faster

## Performance Impact
- **Before**: 2.57s (real 2.715s, user 1.347s, sys 1.449s)
- **After**: 0.47s (real 0.626s, user 0.505s, sys 0.200s)
- **Improvement**: ~82% faster execution time

## Checklist

- [x] Existing or new tests cover this change
- [x] There is a changeset for this change, or one is not required
- [x] Added documentation for any new features to the `docs/` folder

<!-- If this change does not require a changeset, uncomment the tag and explain why -->
<!-- [no-changeset]: Only changing build script -->